refactor: clean unused files
