+++ src/components/component-status/component-status.config.js
@@ -1,4 +1,4 @@
-import { includes } from "lodash";
+import includes from "lodash/includes";
 
 export default function (ComponentStatusProvider) {
     "ngInject";
