+++ src/components/component-status/index.js
@@ -1,0 +1,10 @@
+import ComponentStatus from "./component-status.component";
+import componentStatusConfig from "./component-status.config";
+import ComponentStatusProvider from "./component-status.provider";
+
+export default angular
+    .module("oui-doc.components.component-status", [])
+    .component("componentStatus", ComponentStatus)
+    .provider("ComponentStatus", ComponentStatusProvider)
+    .config(componentStatusConfig)
+    .name;
