--- src/components/init/redirectTo.run.js
@@ -1,10 +1,0 @@
-export default function ($rootScope, $state) {
-    "ngInject";
-
-    $rootScope.$on("$stateChangeStart", (event, toState, toParams) => {
-        if (toState.redirectTo) {
-            event.preventDefault();
-            $state.go(toState.redirectTo, toParams, { location: "replace" });
-        }
-    });
-}
